requirements.txt
fastapi[standard]
uvicorn
selenium
python-dotenv
anticaptchaofficial
2captcha-python
pywhatkit
smsmobileapi
psycopg2
requests
BeautifulSoup
bs4